+++ source/mysql/types.d
@@ -1,0 +1,31 @@
+﻿/// Structures for MySQL types not built-in to D/Phobos.
+module mysql.types;
+
+/++
+A simple struct to represent time difference.
+
+D's std.datetime does not have a type that is closely compatible with the MySQL
+interpretation of a time difference, so we define a struct here to hold such
+values.
++/
+struct TimeDiff
+{
+	bool negative;
+	int days;
+	ubyte hours, minutes, seconds;
+}
+
+/++
+A D struct to stand for a TIMESTAMP
+
+It is assumed that insertion of TIMESTAMP values will not be common, since in general,
+such columns are used for recording the time of a row insertion, and are filled in
+automatically by the server. If you want to force a timestamp value in a prepared insert,
+set it into a timestamp struct as an unsigned long in the format YYYYMMDDHHMMSS
+and use that for the approriate parameter. When TIMESTAMPs are retrieved as part of
+a result set it will be as DateTime structs.
++/
+struct Timestamp
+{
+	ulong rep;
+}